requirements.txt
Django
djangorestframework
channels
pillow